Moved contents to the cofig json file
