+++ src/data/website_config.js
@@ -1,0 +1,7 @@
+import webConfig from './website_data.json';
+//import websiteData from https://raw.githubusercontent.com/pmariammargaret/rk-arkgt-website/dev_20241226/src/data/website_data.json 
+
+let websiteData = webConfig;
+
+
+export default websiteData;
